requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
boto3==1.28.70
botocore==1.31.70
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
deprecation==2.1.0
dj-database-url==2.1.0
Django==4.2.7
django-bootstrap5==23.3
django-email-verification==0.3.3
django-filter==23.3
django-storages==1.14.2
django-webpush==0.3.6
frozenlist==1.4.1
gunicorn==21.2.0
http_ece==1.2.0
idna==3.7
jmespath==1.0.1
multidict==6.0.5
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
py-vapid==1.9.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.8.2
pywebpush==2.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
validators==0.22.0
whitenoise==6.5.0
yarl==1.9.4
